Day 01: Refactor using parse_ints function

+++ 2024/day-01/src/structs.rs
@@ -1,5 +1,5 @@
 #[derive(Clone, Debug)]
 pub struct PuzzleInput {
-    pub left: Vec<u32>,
-    pub right: Vec<u32>,
+    pub left: Vec<i64>,
+    pub right: Vec<i64>,
 }
